+++ src/cli.rs
@@ -70,6 +70,10 @@ pub fn build_cli() -> App<'static, 'static> {
                 .visible_aliases(&["parameter", "params", "param", "p"])
                 .about("Work with CloudTruth parameters")
                 .subcommands(vec![
+                    SubCommand::with_name("delete")
+                        .visible_alias("del")
+                        .about("Delete the parameter from the project")
+                        .arg(Arg::with_name("KEY").required(true)),
                     SubCommand::with_name("get")
                         .about("Gets value for parameter in the selected environment")
                         .arg(Arg::with_name("KEY").required(true).index(1)),
@@ -90,7 +94,8 @@ pub fn build_cli() -> App<'static, 'static> {
                             .help("Parameter 'values' output data format")
                         ),
                     SubCommand::with_name("set")
-                        .about("Set a static value in the selected environment for an existing parameter or creates a new one if needed")
+                        .about(concat!("Set a static value in the selected project/environment for ",
+                            "an existing parameter or creates a new one if needed"))
                         .arg(Arg::with_name("KEY").required(true).index(1))
                         .arg(Arg::with_name("VALUE").required(true).index(2)),
                 ]),